[Refactor] : GitHubSearchClient 파일 분리, 폴더 정리
Git: #11
